server/requirements.txt
Flask==3.0.1
Flask-Cors==4.0.0
langchain==0.0.353
langchain-community==0.0.7
langchain-core==0.1.7
langchain-openai==0.0.2
langchainhub==0.1.14
python-dotenv==1.0.0